+++ banyand/internal/storage/storage.go
@@ -119,6 +119,7 @@ type Segment[T TSTable, O any] interface {
 	GetTimeRange() timestamp.TimeRange
 	CreateTSTableIfNotExist(shardID common.ShardID) (T, error)
 	Tables() []T
+	Caches() []*ShardCache
 	Lookup(ctx context.Context, series []*pbv1.Series) (pbv1.SeriesList, error)
 	IndexDB() IndexDB
 }